+++ packages/firebase_auth/CHANGELOG.md
@@ -1,3 +1,33 @@
+## 0.6.5
+
+* Fixing async method `verifyPhoneNumber`, that would never return even in a successful call.
+
+## 0.6.4
+
+* Added support for Github signin and linking Github accounts to existing users.
+
+## 0.6.3
+
+* Add multi app support.
+
+## 0.6.2+1
+
+* Bump Android dependencies to latest.
+
+## 0.6.2
+
+* Add access to user metadata.
+
+## 0.6.1
+
+* Adding support for linkWithTwitterCredential in FirebaseAuth.
+
+## 0.6.0
+
+* Added support for `updatePassword` in `FirebaseUser`.
+* **Breaking Change** Moved `updateEmail` and `updateProfile` to `FirebaseUser`.
+  This brings the `firebase_auth` package inline with other implementations and documentation.
+
 ## 0.5.20
 
 * Replaced usages of guava's: ImmutableList and ImmutableMap with platform
@@ -15,7 +45,7 @@ Collections.unmodifiableList() and Collections.unmodifiableMap().
 
 ## 0.5.17
 
-* Adding support for FirebaseUser.delete. 
+* Adding support for FirebaseUser.delete.
 
 ## 0.5.16